+++ psalm-baseline.xml
@@ -6152,6 +6152,46 @@
       <code><![CDATA[$_REQUEST['goto']]]></code>
     </PossiblyInvalidCast>
   </file>
+  <file src="src/I18n/LanguageManager.php">
+    <DeprecatedMethod>
+      <code><![CDATA[Config::getInstance()]]></code>
+    </DeprecatedMethod>
+    <MixedArgument>
+      <code><![CDATA[$this->config->get('Lang')]]></code>
+      <code><![CDATA[$this->config->getCookie('pma_lang')]]></code>
+    </MixedArgument>
+    <MixedArrayOffset>
+      <code><![CDATA[$langs[$this->config->get('DefaultLang')]]]></code>
+    </MixedArrayOffset>
+    <MixedArrayTypeCoercion>
+      <code><![CDATA[$langs[$this->config->get('DefaultLang')]]]></code>
+    </MixedArrayTypeCoercion>
+    <MixedOperand>
+      <code><![CDATA[$this->config->get('FilterLanguages')]]></code>
+    </MixedOperand>
+    <PossiblyInvalidArgument>
+      <code><![CDATA[$_GET['lang']]]></code>
+      <code><![CDATA[$_POST['lang']]]></code>
+    </PossiblyInvalidArgument>
+    <PossiblyInvalidCast>
+      <code><![CDATA[$_GET['lang']]]></code>
+      <code><![CDATA[$_POST['lang']]]></code>
+    </PossiblyInvalidCast>
+    <PossiblyNullArgument>
+      <code><![CDATA[$this->config->get('Lang')]]></code>
+      <code><![CDATA[$this->config->getCookie('pma_lang')]]></code>
+    </PossiblyNullArgument>
+    <PossiblyNullArrayOffset>
+      <code><![CDATA[$langs]]></code>
+    </PossiblyNullArrayOffset>
+    <RiskyTruthyFalsyComparison>
+      <code><![CDATA[empty($_GET['lang'])]]></code>
+      <code><![CDATA[empty($_POST['lang'])]]></code>
+      <code><![CDATA[empty($this->config->get('FilterLanguages'))]]></code>
+      <code><![CDATA[empty($this->config->get('Lang'))]]></code>
+      <code><![CDATA[empty($this->config->getCookie('pma_lang'))]]></code>
+    </RiskyTruthyFalsyComparison>
+  </file>
   <file src="src/Image/ImageWrapper.php">
     <PossiblyUnusedReturnValue>
       <code><![CDATA[bool]]></code>
@@ -6422,46 +6462,6 @@
       <code><![CDATA[$ruleData[2]]]></code>
       <code><![CDATA[$ruleData[2]]]></code>
     </PossiblyUndefinedArrayOffset>
-  </file>
-  <file src="src/LanguageManager.php">
-    <DeprecatedMethod>
-      <code><![CDATA[Config::getInstance()]]></code>
-    </DeprecatedMethod>
-    <MixedArgument>
-      <code><![CDATA[$this->config->get('Lang')]]></code>
-      <code><![CDATA[$this->config->getCookie('pma_lang')]]></code>
-    </MixedArgument>
-    <MixedArrayOffset>
-      <code><![CDATA[$langs[$this->config->get('DefaultLang')]]]></code>
-    </MixedArrayOffset>
-    <MixedArrayTypeCoercion>
-      <code><![CDATA[$langs[$this->config->get('DefaultLang')]]]></code>
-    </MixedArrayTypeCoercion>
-    <MixedOperand>
-      <code><![CDATA[$this->config->get('FilterLanguages')]]></code>
-    </MixedOperand>
-    <PossiblyInvalidArgument>
-      <code><![CDATA[$_GET['lang']]]></code>
-      <code><![CDATA[$_POST['lang']]]></code>
-    </PossiblyInvalidArgument>
-    <PossiblyInvalidCast>
-      <code><![CDATA[$_GET['lang']]]></code>
-      <code><![CDATA[$_POST['lang']]]></code>
-    </PossiblyInvalidCast>
-    <PossiblyNullArgument>
-      <code><![CDATA[$this->config->get('Lang')]]></code>
-      <code><![CDATA[$this->config->getCookie('pma_lang')]]></code>
-    </PossiblyNullArgument>
-    <PossiblyNullArrayOffset>
-      <code><![CDATA[$langs]]></code>
-    </PossiblyNullArrayOffset>
-    <RiskyTruthyFalsyComparison>
-      <code><![CDATA[empty($_GET['lang'])]]></code>
-      <code><![CDATA[empty($_POST['lang'])]]></code>
-      <code><![CDATA[empty($this->config->get('FilterLanguages'))]]></code>
-      <code><![CDATA[empty($this->config->get('Lang'))]]></code>
-      <code><![CDATA[empty($this->config->getCookie('pma_lang'))]]></code>
-    </RiskyTruthyFalsyComparison>
   </file>
   <file src="src/Linter.php">
     <InvalidOperand>
@@ -13409,6 +13409,20 @@
       <code><![CDATA[providerForTestGetRoute]]></code>
     </PossiblyUnusedMethod>
   </file>
+  <file src="tests/unit/I18n/LanguageTest.php">
+    <DeprecatedMethod>
+      <code><![CDATA[Config::getInstance()]]></code>
+      <code><![CDATA[Config::getInstance()]]></code>
+      <code><![CDATA[Config::getInstance()]]></code>
+      <code><![CDATA[Config::getInstance()]]></code>
+      <code><![CDATA[Config::getInstance()]]></code>
+      <code><![CDATA[Config::getInstance()]]></code>
+    </DeprecatedMethod>
+    <PossiblyUnusedMethod>
+      <code><![CDATA[listLocales]]></code>
+      <code><![CDATA[selectDataProvider]]></code>
+    </PossiblyUnusedMethod>
+  </file>
   <file src="tests/unit/Identifiers/DatabaseNameTest.php">
     <PossiblyUnusedMethod>
       <code><![CDATA[providerForTestInvalidNames]]></code>
@@ -13570,20 +13584,6 @@
     </PossiblyNullArrayOffset>
     <PossiblyUnusedMethod>
       <code><![CDATA[proxyIPs]]></code>
-    </PossiblyUnusedMethod>
-  </file>
-  <file src="tests/unit/LanguageTest.php">
-    <DeprecatedMethod>
-      <code><![CDATA[Config::getInstance()]]></code>
-      <code><![CDATA[Config::getInstance()]]></code>
-      <code><![CDATA[Config::getInstance()]]></code>
-      <code><![CDATA[Config::getInstance()]]></code>
-      <code><![CDATA[Config::getInstance()]]></code>
-      <code><![CDATA[Config::getInstance()]]></code>
-    </DeprecatedMethod>
-    <PossiblyUnusedMethod>
-      <code><![CDATA[listLocales]]></code>
-      <code><![CDATA[selectDataProvider]]></code>
     </PossiblyUnusedMethod>
   </file>
   <file src="tests/unit/LinterTest.php">
