+++ src/I18n/Language.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpMyAdmin;
+namespace PhpMyAdmin\I18n;
 
 use function __;
 use function _bindtextdomain;
